finish crud users

+++ udemy/nestjs-complete-developers-guide/mycarvalue/src/users/users.service.ts
@@ -34,6 +34,9 @@ export class UsersService {
 
   async remove(id: number) {
     const user = await this.findOne(id);
+    if (!user) {
+      return true;
+    }
     return this.repo.remove(user);
   }
 }
